refactor: move getStageColorClasses to shared utils
